+++ liquidhaskell-boot/src/Language/Haskell/Liquid/Constraint/Generate.hs
@@ -1445,7 +1445,7 @@ lamExpr :: CGEnv -> CoreExpr -> CG (Maybe F.Expr)
 lamExpr g e = do
     adts <- gets cgADTs
     let dm = dataConMap adts
-    return $ eitherToMaybe $ runToLogic (emb g) mempty dm (getConfig g)
+    return $ eitherToMaybe $ runToLogic (emb g) mempty (Just dm) (getConfig g)
       (\x -> todo Nothing ("coreToLogic not working lamExpr: " ++ x))
       (coreToLogic e)